requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
typing-extensions==4.8.0
requests==2.31.0
aiohttp==3.9.1
llama-cpp-python==0.2.20 